Implement & test `util::IDEServices::showMessage`.
