+++ rascal-lsp/src/main/java/org/rascalmpl/vscode/lsp/LSPIDEServices.java
@@ -42,8 +42,10 @@ import org.rascalmpl.ideservices.IDEServices;
 import org.rascalmpl.uri.URIUtil;
 import org.rascalmpl.vscode.lsp.rascal.conversion.Diagnostics;
 import org.rascalmpl.vscode.lsp.rascal.conversion.DocumentChanges;
+import org.rascalmpl.vscode.lsp.rascal.conversion.Message;
 import org.rascalmpl.vscode.lsp.util.locations.Locations;
 
+import io.usethesource.vallang.IConstructor;
 import io.usethesource.vallang.IInteger;
 import io.usethesource.vallang.IList;
 import io.usethesource.vallang.IMap;
@@ -177,6 +179,11 @@ public class LSPIDEServices implements IDEServices {
 
     public IRascalMonitor getMonitor() {
         return monitor;
+    }
+
+    @Override
+    public void showMessage(IConstructor message) {
+        languageClient.showMessage(Message.toMessageParams(message));
     }
 
 }
